app_tracking_system/requirements.txt
Django == 3.2.18
drf-yasg==1.20.0
djangorestframework==3.13.1
mysqlclient==2.1.0